Update Header.kt
Created State Variables:
Used state variables to store the user's name and media counts. This allows  the user to update the UI when the values change.
Implement a Dialog for Editing:
Created a dialog that pops up when the edit button is clicked, allowing users to enter new values.

Update the Profile Header:
Uses  the updated values in the ProfileHeader.